# Conflicts:
#   db/schema.rb

+++ db/schema.rb
@@ -234,7 +234,7 @@ ActiveRecord::Schema.define(version: 2021_08_06_150601) do
     t.integer "objective", default: 0
     t.integer "impression_count", default: 0
     t.float "total_invested", default: 0.0
-    t.integer "people_reached", default: 0
+    t.float "people_reached", default: 0.0
     t.integer "duration", default: 10
     t.integer "lock_version"
     t.boolean "updating_state", default: false
